requirements.txt
jsonschema
matrix-nio[e2e]
pyshacl
rdflib
